fix: move all the provider inside providers
