+++ src/eth_rpc/servers/alchemy_rpc.rs
@@ -1,9 +1,9 @@
 #![allow(clippy::blocks_in_conditions)]
 
 use crate::{
-    alchemy_provider::provider::AlchemyProvider,
     eth_rpc::api::alchemy_api::AlchemyApiServer,
     models::token::{TokenBalances, TokenMetadata},
+    providers::alchemy_provider::provider::AlchemyProvider,
 };
 use async_trait::async_trait;
 use jsonrpsee::core::RpcResult as Result;
