+++ src/lib.rs
@@ -4,13 +4,15 @@ use opentelemetry_sdk as _;
 use tracing_opentelemetry as _;
 use tracing_subscriber as _;
 
-pub mod alchemy_provider;
+pub mod providers {
+    pub mod alchemy_provider;
+    pub mod debug_provider;
+    pub mod eth_provider;
+    pub mod pool_provider;
+}
 pub mod config;
-pub mod debug_provider;
-pub mod eth_provider;
 pub mod eth_rpc;
 pub mod models;
-pub mod pool_provider;
 pub mod prometheus_handler;
 pub mod retry;
 #[cfg(feature = "testing")]
