+++ src/providers/eth_provider/database/types/header.rs
@@ -11,7 +11,7 @@ use {
 #[derive(Debug, Serialize, Deserialize, Hash, Clone, PartialEq, Eq)]
 #[cfg_attr(any(test, feature = "arbitrary", feature = "testing"), derive(arbitrary::Arbitrary))]
 pub struct StoredHeader {
-    #[serde(deserialize_with = "crate::eth_provider::database::types::serde::deserialize_intermediate")]
+    #[serde(deserialize_with = "crate::providers::eth_provider::database::types::serde::deserialize_intermediate")]
     pub header: Header,
 }
