+++ src/providers/eth_provider/database/types/receipt.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 /// A transaction receipt as stored in the database
 #[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone)]
 pub struct StoredTransactionReceipt {
-    #[serde(deserialize_with = "crate::eth_provider::database::types::serde::deserialize_intermediate")]
+    #[serde(deserialize_with = "crate::providers::eth_provider::database::types::serde::deserialize_intermediate")]
     pub receipt: TransactionReceipt,
 }
