+++ src/providers/eth_provider/database/types/transaction.rs
@@ -18,7 +18,7 @@ use {
 #[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone)]
 #[cfg_attr(any(test, feature = "arbitrary", feature = "testing"), derive(arbitrary::Arbitrary))]
 pub struct StoredTransaction {
-    #[serde(deserialize_with = "crate::eth_provider::database::types::serde::deserialize_intermediate")]
+    #[serde(deserialize_with = "crate::providers::eth_provider::database::types::serde::deserialize_intermediate")]
     pub tx: Transaction,
 }
 
@@ -166,7 +166,7 @@ impl<'a> StoredTransaction {
 #[derive(Debug, Serialize, Deserialize, Eq, PartialEq, Clone)]
 pub struct StoredPendingTransaction {
     /// Transaction object
-    #[serde(deserialize_with = "crate::eth_provider::database::types::serde::deserialize_intermediate")]
+    #[serde(deserialize_with = "crate::providers::eth_provider::database::types::serde::deserialize_intermediate")]
     pub tx: Transaction,
     /// Number of retries
     pub retries: u8,
