+++ src/providers/eth_provider/error.rs
